+++ apps/chat/src/components/Chat/MemoizedChatMessage.tsx
@@ -1,5 +1,5 @@
 import { FC, memo } from 'react';
 
-import { ChatMessage, Props } from './ChatMessage';
+import { ChatMessage, Props } from './ChatMessage/ChatMessage';
 
 export const MemoizedChatMessage: FC<Props> = memo(ChatMessage);
