+++ apps/chat/src/components/Common/DropdownMenu.tsx
@@ -67,9 +67,11 @@ interface MenuProps {
   nested?: boolean;
   children?: ReactNode;
   type?: 'dropdown' | 'contextMenu';
-  placement?: Placement;
   isMenuOpen?: boolean;
   onOpenChange?: (isOpen: boolean) => void;
+  placement?: Placement;
+  shouldFlip?: boolean;
+  shouldApplySize?: boolean;
 }
 
 export const MenuComponent = forwardRef<
@@ -87,6 +89,8 @@ export const MenuComponent = forwardRef<
     placement,
     isMenuOpen,
     onOpenChange,
+    shouldFlip = true,
+    shouldApplySize = true,
     ...props
   },
   forwardedRef,
@@ -121,17 +125,21 @@ export const MenuComponent = forwardRef<
     placement: placement ?? (isNested ? 'right-start' : 'bottom-start'),
     middleware: [
       offset(0),
-      flip(),
+      ...(shouldFlip ? [flip()] : []),
       shift(),
-      size({
-        apply({ rects, availableWidth, availableHeight, elements }) {
-          setFloatingWidth(rects.reference.width);
-          Object.assign(elements.floating.style, {
-            maxWidth: `${availableWidth}px`,
-            maxHeight: `${availableHeight}px`,
-          });
-        },
-      }),
+      ...(shouldApplySize
+        ? [
+            size({
+              apply({ rects, availableWidth, availableHeight, elements }) {
+                setFloatingWidth(rects.reference.width);
+                Object.assign(elements.floating.style, {
+                  maxWidth: `${availableWidth}px`,
+                  maxHeight: `${availableHeight}px`,
+                });
+              },
+            }),
+          ]
+        : []),
     ],
     whileElementsMounted: autoUpdate,
   });
